btf: move Handle into handle.go
No other source code changes intended.
